scripts/requirements.txt
# Python: 3.10.16                                                                                                                                                                                    
numpy==2.2.4
psycopg-binary==3.2.9
matplotlib==3.10.1
